+++ examples/components/datagrid/datagrid.component.jsx
@@ -1,27 +1,44 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { List, Map } from 'immutable';
+import ImmutablePropTypes from 'react-immutable-proptypes';
+
 import { Datagrid, DatagridActions } from '../../../src/index';
 import { bankAccountData } from './data';
 import './datagrid.component.scss';
 
-// TODO: Expand example to have inline edit etc..
-
 const GRID_ID = 'accounts-grid-example';
 
 const mapDispatchToProps = {
+  datagridCellShowMessage: DatagridActions.cellShowMessage,
   datagridSetData: DatagridActions.setData,
 };
 
-const mapStateToProps = () => ({});
+const mapStateToProps = state => ({
+  data: state.datagrid.getIn([GRID_ID, 'data'], List()),
+  dataEdited: state.datagrid.getIn([GRID_ID, 'editData'], Map()),
+  isEditing: state.datagrid.getIn([GRID_ID, 'session', 'isEditing'], false),
+});
 
 @connect(mapStateToProps, mapDispatchToProps)
 export default class DatagridView extends React.Component {
   static propTypes = {
+    data: ImmutablePropTypes.list.isRequired,
+    dataEdited: ImmutablePropTypes.map.isRequired,
+    datagridCellShowMessage: PropTypes.func.isRequired,
     datagridSetData: PropTypes.func.isRequired,
+    isEditing: PropTypes.bool.isRequired,
   };
 
   componentWillMount() {
     this.props.datagridSetData(GRID_ID, bankAccountData);
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.data.size !== this.props.data.size) {
+      this.props.datagridCellShowMessage(GRID_ID, 'warning', 1, ['name'], 'WarningExample');
+    }
   }
 
   render() {
@@ -31,6 +48,9 @@ export default class DatagridView extends React.Component {
         valueKeyPath: ['name'],
         valueType: 'text',
         componentType: 'text',
+        validators: [
+          { unique: true },
+        ],
         width: 200,
       },
       {
@@ -78,9 +98,12 @@ export default class DatagridView extends React.Component {
           id={GRID_ID}
           idKeyPath={['accountId']}
           columns={columns}
-          rowSelect
-          multiSelect
+          disableActionSave={this.props.isEditing && this.props.dataEdited.size === 0}
+          enableArrowNavigation
           filtering
+          inlineEdit
+          multiSelect
+          rowSelect
           rowSelectCheckboxColumn
         />
       </div>